requirements.txt
nudepy
pymysql
requests
BeautifulSoup4
lxml
urllib
